backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# WebSocket
python-socketio==5.10.0

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0

# AI Integration
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# Rate Limiting
slowapi==0.1.9

# CORS
starlette==0.27.0

# Environment
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2

